refactor: make new common module
Make new module for common methods, currently the activity extensions reside there
