+++ settings.gradle
@@ -3,3 +3,4 @@ include ':app'
 include ':feature:pusher'
 include ':library:datasource:pwpush'
 include ':library:resources'
+include ':library:common'
